+++ CHANGES.md
@@ -58,6 +58,7 @@ Changes in 13.0
 - Remove local `_reloadMenu()` function in Modules.inc.php
 - Add RosarioSIS\Functions\AddonInfo class in AddonInfo.php
 - Check add-on requirements in Modules.inc.php & Plugins.inc.php
+- Add AddonMakeUpdateAvailable() function in Addon.fnc.php
 
 Changes in 12.9.2
 -----------------